版本2：(引入对目标代码和状态的双定向)加入在TMP_DIR/state_targets.txt中提取目标状态,修改了aflnet.h中distance_to_target_state错误的u32类型声明

+++ aflnet.h
@@ -32,7 +32,7 @@ typedef struct {
   u32 seeds_count;            /* total number of seeds, it must be equal the size of the seeds array */
   
   //aflnet_go
-  u32 distance_to_target_state;				  /* Distance to target state */
+  double distance_to_target_state;				  /* Distance to target state */
   //aflnet_go#
 } state_info_t;
